requirements.txt
asgiref==3.7.2
certifi==2022.6.15
distlib==0.3.4
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
filelock==3.7.1
gunicorn==21.2.0
packaging==23.1
pipenv==2022.6.7
platformdirs==2.5.2
psycopg2==2.9.6
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.15.1
virtualenv-clone==0.5.7